refactor(modularization): Phase 5 step 5-S4e — chat_title_service + DefaultModelOps DI (task #26)
Move the fifth of the six conversation services into its canonical
domain home and swap the direct import of
``aperag.service.default_model_service`` for a new consumer-owned
``DefaultModelOps`` Protocol DI slot. Canonical ``msg=b93904b6`` +
``msg=3ca9dea3`` rule the DI path in even though Phase 4 has already
placed the concrete provider in ``aperag/domains/model_platform`` —
the Protocol + DI pattern stays consistent with the other Phase 5
consumer-owned seams (``QuotaOps`` / ``ChatDocumentOps`` /
``ChatCollectionServiceOps``). Phase 6 cleanup is the right place to
evaluate whether this whole class of narrow DI-with-stable-provider
slots should collapse into direct imports.

Changes:

* ``aperag/domains/conversation/ports.py`` — new
  ``DefaultModelOps(Protocol)`` with one method,
  ``get_default_background_task_config(user_id) -> Any``, added to
  ``__all__``.
* ``aperag/domains/conversation/service/chat_title_service.py`` —
  moved from ``aperag/service/chat_title_service.py`` with:
  - ``from aperag.service.default_model_service import default_model_service``
    → ``from aperag.domains.conversation.ports import DefaultModelOps``
    + ``_default_model_ops`` slot + ``set_default_model_ops`` setter +
    ``_get_default_model_ops`` accessor (RuntimeError-on-unwired).
  - Call site ``default_model_service.get_default_background_task_config(user_id)``
    → ``_get_default_model_ops().get_default_background_task_config(user_id)``.
* ``aperag/service/chat_title_service.py`` — collapses to a re-export
  shim covering ``ChatTitleService``, ``chat_title_service``, and
  ``set_default_model_ops`` for pre-migration callers
  (``views/chat.py``).
* ``aperag/app.py`` — imports
  ``aperag.domains.model_platform.service.default_model_service``
  (the Phase 4 canonical home, not the legacy shim) and calls
  ``_conv_set_default_model_ops`` at module scope alongside the other
  conversation DI wire-ups.
* ``tests/unit_test/chat/test_chat_title_service.py`` — patch target
  updated from the deleted
  ``aperag.service.chat_title_service.default_model_service`` symbol
  to the canonical
  ``aperag.domains.conversation.service.chat_title_service._default_model_ops``
  slot. A stub ``SimpleNamespace`` with the single method replaces
  the whole slot so the test-shape stays identical.

Verify:
* ``import aperag.app`` then
  ``aperag.domains.conversation.service.chat_title_service._default_model_ops``
  is non-None.
* ``from aperag.service.chat_title_service import ChatTitleService,
  chat_title_service, set_default_model_ops`` OK via shim; class
  identity preserved against the domain canonical home.
* ``pytest tests/unit_test/`` 673 passed / 29 skipped / 1 deselected
  (performance flake pre-existing).
* ``uvx ruff check`` + ``uvx ruff format --check`` clean.
* G1 on ``aperag/domains/conversation/service/chat_title_service.py``
  zero legacy aggregate imports.
